Add files via upload
This topic is "How to create a Treemap".

Treemaps are used to display the data in the form of nested rectangles.
In general dimensions are used to define the structure of the treemap, and measures are used to define the size or color of the individual rectangles. 
Treemaps are a relatively simple data visualization that can provide insight in a visually attractive format.
Treemaps are a better alternative to Pie Charts when the number of members are more.

Treemaps fall under the Composition Visual Segment.

The basic building blocks for a treemap are as follows:
Mark Type: Automatic or Square
Color: Dimension or Measure
Size: Measure
Detail: Dimension(s)

Given below are the basic steps to create a Treemap without using the "Show Me" option:

Step 1: Drag and drop any Measure e.g.: Sales to the Color and Size on the Marks card

Step 2: Drag and drop any Dimension e.g.: Sub-Category to the Detail on the Marks card

Step 3: Drag and drop the same Dimension e.g.: Sub-Category to the Label on the Marks card

Step 4: Drag and drop the same Measure e.g.: Sales to the Label on the Marks card
Phones have the highest Sales with the largest size of rectangle as well as the darkest color    

Step 5: If required we can divide this view further based on another Dimension e.g.: Segment
Drag and drop 2nd Dimension e.g.: Segment to the Color on the Marks card  
Please note that a Color Legend appears in the Top Right side of the view

Step 6: Now the Treemap view is complete
NOTE: There might be few sections that are blank as Tableau does not overlap the labels.
The labels will be visible in Tooltip when we hover over the specific items.
  
The exact steps are mentioned in the attached document.