Add files via upload
This topic is "How to create a Histogram".

Histogram falls under the Distribution Visual Segment.

Histogram are used to display the frequency of distribution or shape of the distribution in the data. 
Histograms look similar to bar charts but differ from them in many ways.
Given below are some of the key differences:
1. A histogram represents the frequency distribution of continuous variables. Conversely, a bar graph is a diagrammatic comparison of discrete variables. 
2. Histogram presents numerical data whereas bar graph shows categorical data. 
3. The histogram is drawn in such a way that there is no gap between the bars. On the other hand, there is proper spacing between bars in a bar graph that indicates discontinuity.
4. In the case of a bar graph, it is quite common to rearrange the blocks, from highest to lowest. But with histogram, this cannot be done, as they are shown in the sequences of classes.

A Histogram in Tableau will typically have a COUNT(Measure) in Rows and Measure(bin) in Columns.
One measure is sufficient to plot a Histogram.
Binning a measure creates a dimension, so we will find newly created bin in the Data Pane

Assume we need to find the distribution of the Quantity field i.e. the number of entries for each quantity 

Given below are the basic steps to create a Histogram without using the "Show Me" option:

Step 1: Drag and drop any measure e.g.: Quantity to the Rows or Columns Shelf

Step 2: In the Rows or Columns Shelf, Right-click the SUM(Quantity) and select Count under the Measure(Sum)
Ensure that the CNT(Quantity) is observed in the Rows or Columns Shelf

Step 3: In the Data Pane, Right-click the measure (Quantity) and select Create --> Bins...

Step 4: Ensure that a new window is observed Edit Bins [Quantity]. Select OK

Step 5: Ensure that a new dimension Quantity(bin) is now present in the Data Pane

Step 6: Drag and drop the Quantity(bin) to the Columns or Rows Shelf

Step 7: In the Columns Shelf, Right-click the Quantity(bin) and select "Continuous" option
Ensure that the Quantity(bin) now do not have any gaps in-between them

Step 8: We can change the bin size by following the below method
In the Data Pane, Right-click Quantity(bin) and select Edit...

Step 9: Change the size of bins to the required number by changing the "Size of bins:" and select OK

Step 10: Ensure that the updated bin size is now reflected in the view
The upper number in each bin is not inclusive. e.g.: 1st bin includes bin sizes 0 to 2 and does not include 3

Step 11: If required the CNT(Quantity) can be displayed by pressing Ctrl and dragging & dropping the CNT(Quantity) to Label section of the Marks card

The exact steps are mentioned in the attached document.