Add files via upload
This topic is "Understand dimensions and measures"

When we connect to a new data source, Tableau performs two actions.
Action#1: Assigns a role (Dimension or Measure)
Action#2: Assigns a type (data type i.e., Number (decimal), Number (whole), Date, Date & Time, String, Boolean, Geospatial)
The role assigned can be either be a dimension or a measure in the Data pane, depending on the type of data the field contains. 

The below mentioned topics will be covered:
Data field roles and types
Using Dimension fields in the view
How dimensions affect the level of detail/granularity of the view
Using Measure fields in the view

The complete details are mentioned in the attached document.