Add files via upload
This Tableau topic is "Understanding Aggregations"

In Tableau, we can aggregate measures or dimensions, though it is more common to aggregate measures. 
Whenever a measure is added to a view, an aggregation is applied to that measure by default. 
The type of aggregation applied varies depending on the context of the view.

Given below are the topics that will be covered:
Change the Aggregation of a Measure in the View
Aggregating Dimensions
Set the Default Aggregation for a Measure
How to Disaggregate Data
How to Disaggregate Data in Scatter Plot
How Aggregated Data changes when Dimensions are added

The attached document has the complete details.