Add files via upload
This Tableau topic is "Understanding Basic Joins"

Joins combine tables by adding more columns of data across similar row structures. 
This can cause data loss or duplication if tables are at different levels of detail, and joined data sources must be fixed before analysis can begin.

When joining tables, the fields that are joined on must be of the same data type. 
If we change the data type after performing the join on the tables, the join will break.

Fields used in the join clause cannot be removed without breaking the join

Data blending simulates a traditional left join. 
The main difference between a Join and Blend is when the aggregation is performed. 
A join combines the data at a row level and then aggregates. 
A blend aggregates and then combines the data.

In most scenarios, traditional table joins are the best choice as they tend to be most performant as the work is done by the database rather than the local machine. 
Table joins are better when tables have a 1:1 relationship (i.e. there is only one record for each value in the linking fields in each table). 
If the tables have a 1: many or many: many relationship this creates duplicated data after the join
Duplicated data can be resolved after the table join is created, but the increased amount of data may cause performance issues.

Given below are the important points w.r.t Joins:

1. The visual cue for a join is a Venn Diagram
2. We need at least 1 data sources to create a join
3. We can have multiple connections while creating joins
4. The structure created during a join is a new combined form 
5. Tables are combined via a Physical join
6. Numbers are aggregated based on the join level granularity

The attached document has the complete details.