Add files via upload
This Tableau topic is "Metadata Grid"

"Metadata grid" provides the data about the data. 

The metadata grid displays the fields in the data source as rows.
This will allow us to quickly examine the structure of the Tableau data source.
After examining the data source we can make use of Metadata grid for renaming fields or hiding multiple fields at once.

NOTE: In Tableau versions prior to 2021.3 we could make use of the the metadata grid button to navigate to the metadata grid.
From Tableau version 2021.3 the view for Data and Metadata grid has changed.

The attached document has the complete details.