Add files via upload
This Tableau Topic is "How to Use Automatic & Custom Split"

We might have some string fields in the data source that contain multiple units of information.
For example, consider we have a full Customer Name.
It might be easier to analyze the data if we split the values in that field into separate fields i.e. First Name and Last Name. 
We can use split or custom split options in Tableau to separate the values based on a separator or a repeated pattern of values present in each row of the field. 

The new fields created by the split or custom split are added to the data source as calculated fields.
Splits and custom splits are based on the SPLIT string function.

Split fields automatically:

A string field can be split automatically based on a common separator that Tableau detects in the field.

Depending on the connection type, a split can automatically separate a field’s values in up to ten new fields. 
The data type of the new fields generated by the split can vary depending on the pattern combination that Tableau detects in the separator that is used to split the original field.


Custom split:

The custom split option can be used to specify a common separator for the split. 
Like the split option, a custom split can separate a field’s values in up to ten new fields. 
In addition, we can choose to split the values at the first n occurrences of the separator, 
the last n occurrences of the separator, or at all occurrences of the separator. 
The data type of the new fields generated by the custom split always results in a string data type.