Add files via upload
This Tableau Topic is "How to utilize Tableau-generated fields"

After we move any data source within Tableau the Data pane will contain an additional number of fields that were not present in the original data. 
These are the Tableau-Generated fields.

Measure Values:
The Measure Values field contains all the measures in the data,
combined into a single field with continuous values.
Measure Values is a generated field containing the values of each measure.
We need to use it along with Measure Names to blend measures.

Measure Names:
The Measure Names field contains the names of all measures in your data, collected into a single field with discrete values.
Measure Names is a generated field containing the names of each measure.
We need to use it along with Measure Values to blend measures.

Count of Table/ Number of Records:
Starting with Tableau 2020.2, every table in a data source has a Count field, in the form of NameofTable(Count).
The table count field is an automatically generated, calculated field.
COUNT of table = SUM of the number of records per table
We can't build calculations on top of a table's Count field, and it is aggregate-only.

Latitude and Longitude:
When Tableau interprets fields to be geographic fields that can be used with maps,
it automatically geocodes the data and includes Latitude (generated) and Longitude fields.
We can use these fields to overlay the data on live maps.

The attached document has the complete details.