Add files via upload
This Tableau Topic is "Understanding Ratio Calculations"

This is an often misunderstood or neglected topic.

It is first important to understand the way Tableau calculates the ratio. It depends on the Aggregation of the data. Therefore, it depends on the granularity or level of details of the view/viz.

Consider we have 2 Calculated Fields:

Profit/Sales : [Profit]/[Sales]
SUM(Profit)/SUM(Sales): SUM([Profit])/SUM([Sales])

For Profit/Sales calculated field, Tableau aggregates the Profit/Sales calculation at the row level and then sums the ratio themselves which is incorrect. Profit/Sales Ratio is always being calculated at the row level or record level and these ratios have been summed.

SUM(Profit)/SUM(Sales) however first sums the Profit and Sales independently at the granularity of the view and then takes the ratio of the two numbers. So irrespective of the granularity of the view the SUM(Profit)/SUM(Sales) calculation is always correct as the aggregation is explicitly mentioned in the calculation itself.

In summary SUM(Profit)/SUM(Sales) sums the Profit and Sales independently to whatever the granularity of the view is and then computes the Ratio at that aggregation. This calculation will produce the correct result irrespective of the granularity of the view.

Profit/Sales always computes the Profit Ratio at the lowest level of granularity and then sums the ratios to the requested aggregation of the view. This calculation will produce huge numbers that will lead to incorrect analysis.

The attached document has the complete details.