Add files via upload
This Tableau Topic is "Working with aggregation options"

Aggregate functions allow us to summarize or change the granularity of the data.

For example, you might want to know exactly how many products were sold in a region etc. 
For this we can use the COUNTD function to summarize the exact number of products sold , and then break the visualization down by region.

The calculation might look something like this:
COUNTD(Product ID)

There are many Aggregate functions available in Tableau but we will be going through only 2 functions
CORR
COUNTD

The examples mentioned are just for explaining the concepts and are totally fictitious.

The attached document has the complete details.