Add files via upload
This Tableau Topic is "Building arithmetic calculations"

Number functions allow us to perform computations on the data values in the fields. Number functions can only be used with fields that contain numerical values. 

For example, you might have a requirement to Rounds a number or a field to the nearest integer of equal or lesser value.
For this we can use the FLOOR function 

The calculation might look something like this:
FLOOR(SUM([Sales]))

There are many Aggregate functions available in Tableau but we will be going through only 2 functions
FLOOR
CEILING

The examples mentioned are just for explaining the concepts and are totally fictitious.

The attached document has the complete details.