Add files via upload
This Tableau Topic is "Using calculations in join clauses"

There might be some instances when the common or shared fields do not create an automatic join. In those cases, a join is performed by setting up one or more join clauses.
The join clause tells Tableau which fields are shared between the tables and how to match the corresponding rows. For example, rows with the same ID are aligned in the results table.

Join clauses most often use the equality operator (=) which matches rows with the same values. It is also possible to perform non-equi joins, such as less than (<) and not equal (<>).
Join clauses can also contain calculations. For example, the join clause could be the concatenation of the name fields “[First name] + [Last name] = [First name] + [Last name]”. Note that not all data source connections support calculations in join clauses.

We will be making use of a fictitious spreadsheet for explaining this concept.

The attached document has the complete details.