Add files via upload
Tableau Topic: Details of the types of Data Sources in Tableau

If a data connection has been created that needs to be used with other workbooks or to be shared with colleagues, we can export (save) the data source to a file. 
The data connection will also include added joined tables, default properties, or custom fields—such as groups, sets, calculated fields, and binned fields—to the Data pane.

Note: This topic describes how to save a data source that we reuse locally, if we do not publish to Tableau Server or Tableau Online. 

The attached document has the complete details.

Tableau Help Link:
https://help.tableau.com/current/pro/desktop/en-us/export_connection.htm