Add files via upload
Tableau Topic: Creating Custom Dates

There might be a requirement to create a custom date if we always use a date at a specific level, or if we want to create calculations that rely on a binned or truncated date, or we might need a different sequence of date parts e.g.: Months and then Years for our analysis.

We may also need to create a special viz like a Cycle plot with only Months and Years Date Parts. These Cycle Plots provide the details regarding the seasonality and trend of data that may not be captured if we use the Dates in the standard chronological order i.e., Years > Quarter > Month etc.

We can create a custom date by working in the Data pane.

The attached document has the complete details.