Add files via upload
Tableau Topic is "View/Viz Level of Detail (or) Granularity"

Before moving onto Level Of Detail expressions it is very important that the View Level Of Detail concept is understood.

Given below are the important points with respect to View/Viz Level of Detail or Granularity

Point#1: It is controlled by Rows , Columns shelf and Color Marks card
Point#2: It is primarily controlled by the Dimensions

NOTE: The other sections of the Marks card can also be used e.g.: Size, Label and Detail. But the change in the Viz is visually more noticeable if we make use of Color

NOTE: There will be no change in granularity if the dimension is moved to the Tooltip of Marks card

In some instances, or use cases we might need to perform calculations or analysis that are not at the View Level Of Detail
This is where we need to make of Level Of Detail Expressions (LOD)

The attached document has the complete details.
The standard sample superstore data set is used for explaining this concept.