Add files via upload
Tableau Quick Tip: Understanding Rows and Columns

It is very important to understand about Rows and Columns in Tableau for better visualization of the data. It will also avoid any assumptions or guesses w.r.t viz orientation

Rows and Columns are called "Shelves" where we can place data fields discrete (blue) or continuous (green)

We can drag fields from the Data pane into these Shelves to create the structure for the visualizations

The icons next to “Rows” and “Columns” will always help us in better understanding the orientation of the visual

The attached document has the complete details.