Add files via upload
Tableau Topic: Multiple Map Layers 

This was a feature that was introduced in Tableau 2020.4 for boosting geospatial analysis. 
With this new multiple marks layers feature we can add an unlimited number of layers to the map. 
This means we can visualize multiple sets of location data in context of one another, and there’s no need for external tools to build custom background maps. 

This allows us to overlay multiple marks layers on a map with no hard limit to the number of layers that can be used.

We will make use of Superstore data set to create three map layers

Layer 1 is for State using a Map mark type where Sales is used as the measure
Layer 2 is for City using a Circle mark type where Quantity is used as the measure
Layer 3 is for Postcode using a Shape mark type where Profit is used as the measure

The attached document has the complete details.