Add files via upload
Tableau Topic: Effective Usage of Show Me

Show Me creates a view based on the fields that are already used in the view and any fields that are selected in the Data pane. 

To use Show Me, select the fields that need to be analyzed in the Data pane, and then select the type of view to be created in the Show Me pane.

Tableau automatically evaluates the selected fields and gives the option of several types of views (non-grayed view types) that would be appropriate for those fields. 

Although there are totally 24 types of visuals that we can plot using Show Me, we will try to create visuals only for 5 examples.

We will follow the below mentioned steps:
Step 1: Create the visual by using the Show Me option
Step 2: Make note of the fields present in Shelves, Marks card and Axes for understanding how Show Me created the first visual
Step 3: We will try to recreate the same visual manually without the Show Me option

The attached document contains the complete details.