Add files via upload
Tableau Topic: How to add a Reference Distribution

"Reference distributions" add a gradient of shading to indicate the distribution of values along the axis. 
Distribution can be defined by percentages, percentiles, quantiles, or standard deviation.

Reference distributions can also be used to create bullet charts.

In this example we will only be going through an example for adding a Reference Distribution to a Bar Chart view.

The exact steps are mentioned in the attached document.