Add files via upload
Tableau Topic: Forecasting and How to create a Forecast

Forecasting in Tableau uses a technique known as exponential smoothing. 
Forecast algorithms try to find a regular pattern in measures that can be continued into the future.

Typically we need a minimum of 1 date dimension, and 1 measure to create a forecast.
However, in the absence of a date, Tableau can create a forecast for a view that contains a dimension with integer values in addition to at least one measure

In this example we will only be going through an example for creating a Forecast for a simple line chart.