Add files via upload
Tableau Topic: Pill that defines the Pane both in Rows and Columns Shelf

I know this might be a very simple topic for many analysts/developers/data folks who have been working on Tableau for many years. But this definition of Pane was something that was not very intuitive for me. Hence I wanted to share it with the data community.

By default, it is always the second last pill in the Rows or Columns Shelf as applicable (based on the Viz Layout) that will define the Pane.

We will go through a simple table calculation (Percent of Total) in a cross-tabs visual. As this way it will be easy to check the Pane definition using the Edit Table Calculation option.

The attached document has the complete details.