Add files via upload
Tableau Topic: Nested Dimensions

For the most basic requirement we will have a view consisting of the below mentioned items
Rows or Columns shelf: One dimension
Rows or Columns shelf: One measure
Label on Marks card: One measure (optional)

But for certain requirements we might need to involve multiple dimensions either in Rows or Columns shelf. In this case we can say that we are having Nested Dimensions in the view.

In general, if there are N dimensions in a shelf the Nesting Level will be "N -1“

We will go though a simple example using the standard Sample Superstore data set.

The attached document has the complete details.