Add files via upload
Tableau Topic: Behavior of using Measure Names and Measure Values on Color in the Marks card & How to assign a separate color palette to each measure when making use of Measure Values in the Color Marks card

Measure Names field contains only the names of all the measures in the data source, combined into a single field with discrete values. 
Since Measure Names is a field with discrete values it will result in a categorical color palette

Measure Values field contains all the measures in the data source, combined into a single field with continuous values. 
Since Measure Values is a field with continuous values it will result in a quantitative color palette
When using Measure Values on Color in case we need to assign a separate color palette for each measure we can make use of the Use Separate Legends option

The attached document has the complete details.