Add files via upload
Tableau Topic: Modify locations within Tableau

There may be certain instances when Tableau may not recognize few locations in our data. Those values show up as unknown in the lower right corner of the map view.

There may be different reasons for this issue:
1. The location or abbreviation may not be recognized by Tableau
2. If there is an ambiguity in location and could be present in multiple places e.g: Auburn is the name of a city in multiple states Alabama, New York and Washington
  
One city name may be present in multiple states. The issue is that Tableau will not know which State to plot to this City into. Hence, we get the Ambiguous red error.

For the unknown entries w.r.t geo fields we can carry out the below mentioned actions:
1. Edit Locations (Correct the unknown locations)
2. Filter data (Exclude the special values from the view and calculations)
3. Show data at default position (Show the special values at a default position on the axis.
For example, Null values are shown at 0)

In this example we will look at the Edit Locations options

We will make use of the standard Sample Super Store data set for this example

The exact steps are mentioned in the attached document.