Add files via upload
Tableau Quick Topic: How do we move the Headers/Labels to the Top of a Chart?

By default the headers will be at the bottom of the chart. It might be possible that the end user might even miss to see the headers/labels at the bottom. If we move the Headers/ Labels to the Top we can avoid the need to scroll down in case of a bigger view to see them.

Go to Analysis > Table Layout > Advanced > Un-Select the "Show innermost level at bottom of view when there is a vertical axis" option

NOTE: This is possible only for discrete fields that includes even discrete dates e.g: MONTH(Order Date) - Blue Pill

The attached document has the complete details.