Add files via upload
Tableau Topic: What are the additional options in discrete filters? (Show All, Show All Values and Show Apply Button)

There are many additional options in discrete filters that we may miss will working on them.

Let us go through a few important options.

Show “All” Value:
This option toggles whether to show the “All” option in the filter

Show All Values Button:
This button toggles whether to display the Show All Values button on the filter card.

Show Apply Button:
This button toggles whether to show the Apply button at the bottom of the filter.
When it is shown, any changes to the filter are only applied after we click the Apply button.
This option is only available in multiple values lists and dropdowns. 


The attached document has the complete details.