Add files via upload
Tableau Topic: How multiple filters behave?

By default, all filters that we set in Tableau are computed independently. That is, each filter accesses all rows in the data source without regard to other filters.
In case there are 2 filters that are applied in a Viz the result will be the intersection of both filters i.e., of all filter values
Like AND operation

We will be using the Sample Superstore data set to check the Multiple Filters behavior.

The attached document has the complete details.