Add files via upload
Tableau Topic: Carrying out Union using Wildcard Search

This method is used to set up search criteria to automatically include tables in the union. Use the wildcard character, which is an asterisk (*), to match a sequence or pattern of characters in the Excel workbook and worksheet names, Google Sheets workbook and worksheet names, text file names, JSON file names, .pdf file names, and database table names.

When working with Excel, text file data, JSON file, .pdf file data, we can also use this method to union files across folders, and worksheets across workbooks. Search is scoped to the selected connection. The connection and the tables available in a connection are shown on the left pane of the Data source page.

The attached document has the complete details.