Add files via upload
Tableau Topic: Data Interpreter

In the ideal scenario we would want to start building our views and dashboards etc. right after connecting to our Excel spreadsheets.

However, in the real-world practical applications it is not a simple step. To make the spreadsheets in a more human readable format we might have added titles, stacked headers, notes, maybe empty rows and columns to add white space (to name a few additions).

Unfortunately, when we want to proceed with analyzing this type of data, these attributes that were added for aesthetic appeal will make it very tough for Tableau to understand the data. This is where the Data Interpreter comes to the rescue.

Data Interpreter can help in performing some initial level of cleaning before starting the actual analysis process. This feature can detect things like titles, notes, footers, empty cells, and so on and bypass them to identify the actual fields and values in the data set.

The document provides the details.