Add files via upload
Tableau Topic: How to Pivot Data in Tableau Desktop?

There are some instances or use cases when it is tougher to analyze the data that is stored in a crosstab format in Tableau.

When working with Microsoft Excel, text file, Google Sheets, and .pdf data sources, we can pivot the data from crosstab format into columnar format. Pivoting data is the technique of data shaping that rotates data from a state of rows to a state of columns. Simply put as the process of converting data from crosstab format (which can be difficulty to work with) to columnar format.

Pivoting data is ideally suited for handling multiple response questions (similar data present across separate fields) and crosstab data

We can access the option by clicking the drop-down arrow next any one of the column names required to be pivoted, Select Pivot

The attached document has the details