Add files via upload
Tableau Topic: Dimension pills are always first placed before the measure pills in either Rows or Columns shelf

If Row or Columns shelves contain both dimensions and measures, Tableau places the measures as the innermost fields, which means that measures are always to the right of any dimensions that we have also placed on these shelves. The word innermost in this case refers to the table structure.

The dimension pills are always first placed before any measure pill in either Rows or Columns shelf. Even if we try to place a dimension after a measure an orange color dropdown carrot indicates that the dimension will be dropped before the measure

Hence Tableau first places all dimensions and finally places the measures (in either Rows or Columns shelf) irrespective of the number of dimensions in the view

The document contains all the details.