Add files via upload
Tableau Topic: Simple method to add multiple measures to a text table

The normal practice to add a measure into a text table is by dragging and dropping it as soon as we get the "Show Me" prompt. This may be tricky sometimes.

The easier method is to just double-click the required measure in the Data pane.

Once we have added 2 measures we can add further measures by either making use of the Measure Names in the Filters shelf or by adding the measures in the Measure Values shelf.


The attached document has the complete details.