Add files via upload
Tableau Topic: How to add a Data Source to the Saved Data Sources?

We might have a requirement to share a data connection with others in a team or to use with other Tableau workbooks. To satisfy this requirement we can carry out an export or save the data source to a file.

Given below are the two formats to save a data source:
Data Source (.tds)
Packaged Data Source (.tdsx)

A .tds file contains only the information you need to connect to the data source, including the following:
Data source type
Connection information specified on the data source page
Groups, sets, calculated fields, bins
Default field properties

A .tdsx file contains all information in the data source (.tds) file, as well as a copy of any local file-based data or extracts.

The document contains all the details of adding a Data Source as a Saved Data Source.