Add files via upload
Tableau Topic: The inner field on Rows or Columns shelves determine the default mark type

Tableau displays data using marks, where every mark corresponds to a row (or a group of rows) in the data source. 
The inner fields on the Rows and Columns shelves determine the default mark type. 
For example, if the inner fields are a measure and a dimension, the default mark type is a Bar. 
We can manually select a different mark type using the Marks card drop-down menu. 

If we place a date field as the inner field the default mark type is a Line.

The attached document contains the details.