Add files via upload
Tableau topic: "How to create a Bar in Bar Chart".

Bar in bar charts are used to compare the measures against the goals, similar to Bullet graphs. 
This chart is actually built on top of a Combined Axis chart.

Bar In Bar Charts like Bar Charts fall under the Comparison Visual Segment.

We will go through the steps to create a Bar in Bar Chart in Tableau.

The exact steps are mentioned in the attached document.