Add files via upload
Tableau Topic: Creating KPI Cards in Tableau

This was one topic that troubled me a bit when I started out in Tableau. Hence this topic is meant for anyone just starting on their Tableau journey.

In Tableau there is no out-of-the-box visual for KPI Cards hence we need to create it manually.
Although there are multiple methods to create KPI Cards in Tableau we will go through only one method.
We will make use of the Measure Values and Measure Names Tableau Generated fields.

The attached document contains the detailed steps to create the KPI Cards.