Add files via upload
Tableau Topic: Replace the connected data source with another data source for an existing chart or sheet
Also includes the mention of the newly added feature in Tableau 2022.4 i.e. Replace datasource per sheet (Replace current worksheet only)

There may be a requirement when we need to update a workbook or sheet to use a different data source.
Instead of building the workbook from scratch using a new data source, we can opt to simply replace the data source.
When we replace a data source does not result in merging or editing the data sources.
Rather, replacing a data source simply redirects fields used in the worksheet to map to corresponding fields in the new data source. 

The attached document has the complete details