Add files via upload
Tableau Topic: The new Image Role

This Image Role was added in Tableau Version 2022.4.

In general images are much better when we need to analyze data. It helps the end-users in better understanding the data insights and the visualizations. 

Image Role is a new field semantic that offers a scalable and automated way to bring image assets into Tableau.
We can now add images to links (URLs) in the data and display them as row or column headers.
This image role feature allows better management of image assets within Tableau and optimize the overall size of workbooks.


The document provides the complete details of how to assign the Image Role from the Data Pane in a Worksheet and from the Data Source Page.