Add files via upload
Tableau topic: "How to create a Line Chart" (without using Show Me)

The Line Chart falls in the Trend Visual Segment.
A Line Chart is typically used to show the trends of a data field over time e.g: Sales between certain year 2019 to 2021 as per a Date Field.
Line charts connect individual data points in a view.
In case we have want to increase the granularity of the line chart we can drop a dimension e.g: Segment to the Color section of Marks card.
The view will change based on the number of members in the dimension that was added. Hence the number of marks in the viz will increase.


The basic building blocks for a Line Chart are as follows:
Mark Type: Line
Rows or Columns Shelf: at least 1 Date
Columns or Rows Shelf: at least 1 Measure
If required we can add additional fields to these shelves
Color: Dimension or Measure (optional)

In case we have want to increase the granularity of the line chart we can drop a dimension e.g: Segment to the Color section of Marks card.
The view will change based on the number of members in the dimension that was added. Hence the number of marks in the viz will increase.

By default when any date is dropped into the Rows/Columns it will be in Discrete Date Part in Years (Blue Color) e.g: YEAR(Order Date)
In this case the Date field will form a Header since it is discrete.

If required we can convert it to Continuous Date Value (Green Color) e.g: MONTH(Order Date)
In this case the Date field will form an Axis.


The attached document provides the exact steps to create a Line Chart in Tableau without using the "Show Me" option.